Use common logic for postgres to geojson
