--- f/export/postgres_to_geojson/tests/requirements-test.txt
@@ -1,2 +1,0 @@
-pytest
-testing.postgresql
